Создание модуляции по файлам, создание dictionary.js, utils.js, data.js
